Move definitions into source
