+++ divi/src/Makefile.am
@@ -172,6 +172,7 @@ BITCOIN_CORE_H = \
   rpcserver.h \
   script/interpreter.h \
   script/InterpreterStack.h \
+  script/StackManager.h \
   script/script.h \
   script/sigcache.h \
   script/sign.h \
@@ -419,6 +420,7 @@ libbitcoin_common_a_SOURCES = \
   pubkey.cpp \
   script/interpreter.cpp \
   script/InterpreterStack.cpp \
+  script/StackManager.cpp \
   script/script.cpp \
   script/sign.cpp \
   script/standard.cpp \
@@ -551,6 +553,7 @@ libbitcoinconsensus_la_SOURCES = \
   pubkey.cpp \
   script/script.cpp \
   script/interpreter.cpp \
+  script/StackManager.cpp \
   script/bitcoinconsensus.cpp \
   uint256.cpp \
   utilstrencodings.cpp
